Uploaded some basic PCB footprints
Upload FPs to have a look at the area constraints in the PCB and see
what sort of area I have to work with.  See the uploaded PCB file for
reference.  This is ONLY for reference.  I will trash the board as the
schematics are completed and I lay it all out.